requirements.txt
Flask
Werkzeug
python-dotenv
flask_sqlalchemy
pymysql
openpyxl
pandas